Adding file recognition

+++ backend/app/src/main/java/com/bls/patronage/resources/DecksCvMagicResource.java
@@ -2,6 +2,7 @@ package com.bls.patronage.resources;
 
 import com.bls.patronage.StorageException;
 import com.bls.patronage.StorageService;
+import com.bls.patronage.api.AcceptableFileTypes;
 import com.bls.patronage.cv.CVRequest;
 import com.bls.patronage.cv.CVResponse;
 import com.bls.patronage.db.dao.DeckDAO;
@@ -16,6 +17,7 @@ import javax.ws.rs.Consumes;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
 import javax.ws.rs.client.Entity;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -43,12 +45,19 @@ public class DecksCvMagicResource {
     @POST
     @Consumes(MediaType.MULTIPART_FORM_DATA)
     public Response uploadFile(@Auth User user,
-                               @FormDataParam("file") InputStream inputStream) throws StorageException {
+                               @FormDataParam("file") InputStream inputStream,
+                               @QueryParam("fileType") AcceptableFileTypes type) throws StorageException {
+
+        if (!type.isValidType())
+            return Response
+                    .status(400)
+                    .entity(Entity.json(type.getGivenType() + " is not a valid file type!"))
+                    .build();
 
         final UUID dataId = storageService.create(inputStream, user.getId());
         final URI publicURLToUploadedFile = storageService.createPublicURL(StorageResource.class, dataId, user.getId());
 
-        final Response flashcards = recoginzeFlashcards(publicURLToUploadedFile);
+        final Response flashcards = recoginzeFlashcards(publicURLToUploadedFile, type.getGivenType());
         save(flashcards, user.getId());
 
         storageService.delete(dataId, user.getId());
@@ -56,10 +65,10 @@ public class DecksCvMagicResource {
                 .status(Response.Status.CREATED).build();
     }
 
-    private Response recoginzeFlashcards(final URI publicURLToUploadedFile) {
+    private Response recoginzeFlashcards(final URI publicURLToUploadedFile, String givenType) {
         return cvServer
                 .request()
-                .buildPost(Entity.json(CVRequest.createRecognizeRequest(publicURLToUploadedFile)))
+                .buildPost(Entity.json(CVRequest.createRecognizeRequest(publicURLToUploadedFile, givenType)))
                 .invoke();
     }
